requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=7.4.0
pymongo>=4.6.0
python-dotenv>=1.0.0